Fix Authors deprecations #49
